メモ保存処理の実装
