+++ app/javascript/packs/application.js
@@ -8,6 +8,7 @@ require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
 require("../memo")
+require("../post")
 import "bootstrap"
 import "../stylesheets/application"
