+++ app/views/memos/index.html.erb
@@ -288,7 +288,7 @@
     <div class="main-text">
       <div class="main-form">
       <%= form_with model: @memo, id: "form" do |f| %>
-        <%= f.text_area :memo, class: "form-control", id: "text-area" %>
+        <%= f.text_area :memo, class: "form-control", id: "textarea" %>
       <% end %>
       </div>
     </div>
